refactor(All-project): Refactor to match package format
