+++ job_scrapper/cli/cli_centraliser.py
@@ -2,9 +2,9 @@ from job_scrapper.cli.scrapper_cli import scrap
 # from sql.cli.enumerate_cli import enumerate
 from job_scrapper.cli.ask_contact import configure_contact
 from cli import cli
-from sql.cli.keyword_cli import manage_keywords
-from sql.cli.request_offer_job_cli import request, request_format_help
-from sql.cli.localisation_cli import geolocalisation
+from job_scrapper.sql.cli.keyword_cli import manage_keywords
+from job_scrapper.sql.cli.request_offer_job_cli import request, request_format_help
+from job_scrapper.sql.cli.localisation_cli import geolocalisation
 from cli import configure
 from job_scrapper.cli.offers_cli import offers
