+++ job_scrapper/scrap_job_from/.template.py
@@ -2,8 +2,7 @@ from bs4 import BeautifulSoup
 from typing import Iterator, Self
 
 import job_scrapper.scrapper_skeleton.scrapper_skeleton as srk
-from web_processing.block_extractor import WebBlockExtractor
-from web_processing.enhanced_chrome_browser import ButtonFinder, EnhancedChrome
+
 
 class TemplateScrapper(srk.JobScrapperSkeleton):
     """
