+++ job_scrapper/scrap_job_from/ifremer.py
@@ -2,7 +2,7 @@ from bs4 import BeautifulSoup
 from typing import Iterator, Self
 
 import job_scrapper.scrapper_skeleton.scrapper_skeleton as srk
-from web_processing.enhanced_chrome_browser import EnhancedChrome
+from job_scrapper.web_processing import EnhancedChrome
 from selenium.webdriver.common.by import By
 
 class IfremerScrapper(srk.JobScrapperSkeleton):
