+++ job_scrapper/scrap_job_from/sfbi.py
@@ -1,17 +1,13 @@
 import re
 
-from bs4 import BeautifulSoup
 from selenium.webdriver.common.by import By
-from selenium.webdriver.remote.webdriver import WebDriver
 from selenium.webdriver.remote.webelement import WebElement
 
-import job_scrapper.scrapper_skeleton.scrapper_skeleton as srk
 from bs4 import BeautifulSoup
 from typing import Iterator, Self
 
 import job_scrapper.scrapper_skeleton.scrapper_skeleton as srk
-from web_processing.block_extractor import WebBlockExtractor
-from web_processing.enhanced_chrome_browser import ButtonFinder, EnhancedChrome
+from job_scrapper.web_processing import EnhancedChrome
 
 class SFBIScrapper(srk.JobScrapperSkeleton):
     """
