+++ job_scrapper/scrapper_skeleton/object_core.py
@@ -2,10 +2,10 @@ import os.path
 import time
 from typing import Sequence
 
-from tools.logger_core import CoreLogger
+from job_scrapper.tools.logger_core import CoreLogger
 import unicodedata
 import re
-from tools.get_unique_path import get_unique_path
+from job_scrapper.tools import get_unique_path
 
 # pylint: disable=R0902
 # This is the amount of  instance attributes that I need
