+++ job_scrapper/scrapper_skeleton/sql_core.py
@@ -8,7 +8,7 @@ from typing import Any, ContextManager, Generator, Protocol, Type, Self, Iterato
 from sqlalchemy.orm import Query, Session
 
 from job_scrapper.scrapper_skeleton.object_core import ScrapperObjectCore
-from sql.tables import (
+from job_scrapper.sql import (
     BaseTable,
     Distances,
     Jobs,
@@ -19,8 +19,8 @@ from sql.tables import (
     TimeStamps,
     Places,
 )
-from sql.tables.helpers.job_request import JobRequest
-from sql.tables.helpers.keyword_manager import KeywordManager
+from job_scrapper.sql.tables.helpers.job_request import JobRequest
+from job_scrapper.sql.tables.helpers import KeywordManager
 
 class SqlSessionFactory(Protocol):
     """Typing class mostly used for ScrapperSQLightCore.get_available_databases"""
