+++ job_scrapper/sql/cli/archive_offer_cli.py
