+++ job_scrapper/sql/cli/request_offer_job_cli.py
