+++ job_scrapper/sql/filters/filter_generator.py
