+++ job_scrapper/sql/tables/README.md
