+++ job_scrapper/sql/tables/__init__.py
@@ -1,0 +1,10 @@
+from job_scrapper.sql.tables.job_extensions.keywords import Keywords
+from job_scrapper.sql.tables.job_extensions.metadata import Metadata
+from job_scrapper.sql.tables.job_extensions.time_stamps import TimeStamps
+from job_scrapper.sql.tables.places.distances import Distances
+from job_scrapper.sql.tables.places.places import Places
+from job_scrapper.sql.tables.keywords.keyword_version import KeywordVersion
+from job_scrapper.sql.tables.keywords.keyword_regex import KeywordRegex
+
+from .base_table import BaseTable
+from .jobs import Jobs
