+++ job_scrapper/sql/tables/base_table.py
