+++ job_scrapper/sql/tables/helpers/README.md
