+++ job_scrapper/sql/tables/job_extensions/README.md
