+++ job_scrapper/sql/tables/job_extensions/metadata.py
