+++ job_scrapper/sql/tables/jobs.py
