+++ job_scrapper/sql/wrappers/wrapper_comparison.py
