+++ job_scrapper/sql/wrappers/wrapper_logical.py
