+++ job_scrapper/tools/logger_core.py
