+++ job_scrapper/tools/variable_tracker.py
