+++ pyproject.toml
@@ -1,6 +1,6 @@
 [project]
-name = "JobScrapper"
-version="0.5.0"
+name = "job-scrapper"
+version="0.6.0"
 description = "A simple library that dowload job offers from several instituts."
 authors = [
     {name="Marhal Florent", email="<marchal.florent.pro@gmail.com>"}
@@ -8,8 +8,12 @@ authors = [
 readme = "README.md"
 license = "CC BY-SA 4.0"
 
+packages = [
+  { include = "job_scrapper" }
+]
+
 [project.scripts]
-cli = "main:main"
+job-scrapper = "job_scrapper.main:main"
 
 [tool.poetry.dependencies]
 python = ">=3.11,<3.12"
