--- sql/tables/__init__.py
@@ -1,10 +1,0 @@
-from sql.tables.job_extensions.keywords import Keywords
-from sql.tables.job_extensions.metadata import Metadata
-from sql.tables.job_extensions.time_stamps import TimeStamps
-from sql.tables.places.distances import Distances
-from sql.tables.places.places import Places
-from sql.tables.keywords.keyword_version import KeywordVersion
-from sql.tables.keywords.keyword_regex import KeywordRegex
-
-from .base_table import BaseTable
-from .jobs import Jobs
